fix(auth): retain refresh compatibility
